+++ src/com/runelive/world/content/LoyaltyProgramme.java
@@ -114,11 +114,11 @@ public class LoyaltyProgramme {
 		player.getPointsHandler().incrementLoyaltyPoints(pts);
 		player.getAchievementAttributes().incrementTotalLoyaltyPointsEarned(pts);
 
-		int totalPoints = (int) player.getPointsHandler().getLoyaltyPoints();
+		int totalPoints = player.getPointsHandler().getLoyaltyPoints();
 
 		if (player.getInterfaceId() == 43000) {
 			player.getPacketSender().sendString(43120, "Your Loyalty Points: " + totalPoints);
 		}
-		player.getPacketSender().sendString(55081, "@red@Loyalty Points: @gre@" + Misc.format((int) totalPoints));
+		player.getPacketSender().sendString(55081, "@red@Loyalty Points: @gre@" + Misc.format(totalPoints));
 	}
 }
