+++ src/com/runelive/world/content/skill/impl/construction/sawmill/SawmillOperator.java
@@ -33,7 +33,7 @@ public class SawmillOperator {
 					.sendMessage("The sawmill operator recognises your cape and gives you the planks for free!");
 		} else {
 			if (player.getMoneyInPouch() >= toPay) {
-				player.setMoneyInPouch(player.getMoneyInPouch() - (long) toPay);
+				player.setMoneyInPouch(player.getMoneyInPouch() - toPay);
 				player.getPacketSender().sendString(8135, "" + player.getMoneyInPouch() + "");
 			} else
 				player.getInventory().delete(995, toPay);
